Rewrote baseEntity
