--- server/src/main/java/net/hamsterapps/cedserver/model/Auditable.java
@@ -1,46 +1,0 @@
-package net.hamsterapps.cedserver.model;
-
-import java.time.Instant;
-
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class Auditable {
-
-  @CreatedDate
-  @Column(name = "created_at", nullable = false)
-  protected Instant createdAt;
-
-  @LastModifiedDate
-  @Column(name = "updated_at", nullable = false)
-  protected Instant updatedAt;
-
-  public Auditable() {
-  }
-
-  public Auditable(Instant createdAt, Instant updatedAt) {
-    this.createdAt = createdAt;
-    this.updatedAt = updatedAt;
-  }
-
-  public Instant getCreatedAt() {
-    return this.createdAt;
-  }
-
-  public Instant getUpdatedAt() {
-    return this.updatedAt;
-  }
-
-  @Override
-  public String toString() {
-    return "{" + "createdAt='" + getCreatedAt() + "'" + ", updatedAt='" + getUpdatedAt() + "'" + "}";
-  }
-
-}
